Pull out opsworks client
