+++ lib/opsicle/aws_instance_manager_helper.rb
@@ -1,32 +1,6 @@
 
 module Opsicle
   module AwsInstanceManagerHelper
-    def select_layer
-      puts "\nLayers:\n"
-      ops_layers = @opsworks.describe_layers({ :stack_id => @stack.id }).layers
-
-      layers = []
-      ops_layers.each do |layer|
-        layers << ManageableLayer.new(layer.name, layer.layer_id, @stack, @opsworks, @ec2, @cli)
-      end
-
-      layers.each_with_index { |layer, index| puts "#{index.to_i + 1}) #{layer.name}" }
-      layer_index = @cli.ask("Layer?\n", Integer) { |q| q.in = 1..layers.length.to_i } - 1
-      layers[layer_index]
-    end
-
-    def select_instances(instances)
-      puts "\nInstances:\n"
-      instances.each_with_index { |instance, index| puts "#{index.to_i + 1}) #{instance.status} - #{instance.hostname}" }
-      instance_indices_string = @cli.ask("Instances? (enter as a comma separated list)\n", String)
-      instance_indices_list = instance_indices_string.split(/,\s*/)
-      instance_indices_list.map! { |instance_index| instance_index.to_i - 1 }
-
-      return_array = []
-      instance_indices_list.each do |index|
-        return_array << instances[index]
-      end
-      return_array
-    end
+    
   end
 end
